refactor: migrate sidebar chapters to react component
